Add abstract base class for image_auto_tag queue workers
